+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   argument is parsed.
 - `FromArgs` derive macro has new option `with` to set custom parser for the
   given field.
+- Add method `read_while` to `Reader`.
 
 ### Changes
 - `cur_manual` and `next_manual` take `FnOnce` instead of `Fn`.